+++ apps/web/src/app/(backButton)/layout.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
+import { PageTitleHeader } from '@/components/header'
 
 export default function BackButtonLayout({ children }: { children: React.ReactNode }) {
     return (
         <div>
-            <div>back button</div>
+            <PageTitleHeader />
             {children}
         </div>
     )
