+++ apps/web/src/components/header/index.ts
@@ -1,0 +1,2 @@
+export * from './logo'
+export * from './pageTitle'
