+++ apps/web/src/components/header/logo.tsx
@@ -1,0 +1,7 @@
+export function LogoHeader() {
+    return (
+        <div className="flex p-2 h-16 border-b-2 items-end">
+            <span className="text-2xl">Fienmee</span>
+        </div>
+    )
+}
